requirements.txt
opencv-python-headless==4.9.0.80
Pillow==10.1.0
numpy==1.24.3
matplotlib==3.7.2